Add files via upload
Uploaded 1726 Real Española Regular Webfont (woff2) to Git Hub to be hosted for World Anvil. I purchased a web license to use this font; please see the license file for further information. If you have not purchase a license to use this file, you are prohibited from using it in any form per the license agreement.

Signed-off-by: Kiel Denny <147275404+kielDenny@users.noreply.github.com>